pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document_analyzer"
version = "0.2.0"
description = "文書分析ツール"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=1.8.2",
    "python-dotenv>=0.19.0",
    "click>=8.0.1",
    "rich>=10.9.0",
    "markdown>=3.3.4",
    "python-multipart>=0.0.20",
    "google-generativeai>=0.1.0",
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "black>=21.8b0",
    "isort>=5.9.3",
    "flake8>=3.9.2",
]

[project.scripts]
document-analyzer = "document_analyzer.cli:main"

[tool.setuptools]
packages = ["document_analyzer"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"